double eye, need to swap upper and lower

+++ Examples/MCUXpresso/i.MX RT1050_EVK/MIMXRT1052_LCD/source/GDisp1.h
@@ -15,7 +15,7 @@
 #define APP_IMG_WIDTH 480
 #define APP_ELCDIF LCDIF
 
-extern uint32_t s_frameBuffer[2][APP_IMG_HEIGHT][APP_IMG_WIDTH];
+extern uint16_t s_frameBuffer[2][APP_IMG_HEIGHT][APP_IMG_WIDTH];
 extern volatile bool s_frameDone;
 
 void GDisp1_PutPixel(unsigned int x, unsigned int y, uint16_t color);
